+++ frontend/src/router/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import ExamplePage from "../page/main/examplePage";
 import SubmitPage from "../page/main/submitPage";
-import LastPage from "../page/finalImg";
+import LastPage from "../page/finallyImg";
 import SecondPage from "../page/selectImg";
 import LoginPage from "../page/login";
